WIP - re-org invoker trans code and impl initial auth op + simple test.
