+++ e2e/eip3074_test.go
@@ -6,6 +6,8 @@ import (
 	"encoding/json"
 	"github.com/0xPolygon/polygon-edge/crypto"
 	"github.com/0xPolygon/polygon-edge/e2e/framework"
+	"github.com/0xPolygon/polygon-edge/helper/hex"
+	"github.com/0xPolygon/polygon-edge/helper/invoker"
 	"github.com/0xPolygon/polygon-edge/helper/tests"
 	"github.com/0xPolygon/polygon-edge/types"
 	"github.com/stretchr/testify/require"
@@ -48,19 +50,19 @@ func getTestArtifact(name string) (art *compiler.Artifact, err error) {
 	return
 }
 
-type ECDSAKey struct {
+type testECDSAKey struct {
 	k *ecdsa.PrivateKey
 }
 
-func (e *ECDSAKey) Address() ethgo.Address {
+func (e *testECDSAKey) Address() ethgo.Address {
 	return ethgo.Address(crypto.PubKeyToAddress(&e.k.PublicKey))
 }
 
-func (e *ECDSAKey) Sign(hash []byte) ([]byte, error) {
+func (e *testECDSAKey) Sign(hash []byte) ([]byte, error) {
 	return crypto.Sign(e.k, hash)
 }
 
-var _ ethgo.Key = &ECDSAKey{}
+var _ ethgo.Key = &testECDSAKey{}
 
 func TestBasicInvoker(t *testing.T) {
 	senderKey, senderAddr := tests.GenerateKeyAndAddr(t)
@@ -92,7 +94,7 @@ func TestBasicInvoker(t *testing.T) {
 	require.NoError(t, err)
 	require.NotNil(t, invokerAddr)
 
-	sk := &ECDSAKey{k: senderKey}
+	sk := &testECDSAKey{k: senderKey}
 	invokerContract := contract.NewContract(invokerAddr, invokerAbi,
 		contract.WithJsonRPC(srv.JSONRPC().Eth()),
 		contract.WithSender(sk),
@@ -103,7 +105,7 @@ func TestBasicInvoker(t *testing.T) {
 	domainSeparator, ok := res["0"].([32]byte)
 	require.True(t, ok)
 
-	tp := TransactionPayload{
+	tp := invoker.TransactionPayload{
 		To:       types.Address{0xde, 0xad, 0xbe, 0xef},
 		Value:    big.NewInt(1000),
 		GasLimit: big.NewInt(1000),
@@ -118,7 +120,7 @@ func TestBasicInvoker(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, checkHash[:], th)
 
-	tps := TransactionPayloads{tp}
+	tps := invoker.TransactionPayloads{tp}
 	res, err = invokerContract.Call("hashPayloads", ethgo.Latest, tps.Payloads())
 	require.NoError(t, err)
 	checkHash, ok = res["0"].([32]byte)
@@ -128,7 +130,7 @@ func TestBasicInvoker(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, checkHash[:], th)
 
-	it := InvokerTransaction{
+	it := invoker.InvokerTransaction{
 		From:     types.Address{0xca, 0xfe, 0xba, 0xbe},
 		Nonce:    big.NewInt(0),
 		Payloads: tps,
@@ -151,11 +153,16 @@ func TestBasicInvoker(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, checkHash[:], th)
 
-	var is InvokerSignature
+	var is invoker.InvokerSignature
 
-	k := &ECDSAKey{k: receiverKey}
+	k := &testECDSAKey{k: receiverKey}
 	err = is.SignCommit(k, checkHash[:], invokerAddr)
 	require.NoError(t, err)
+
+	println(is.String())
+	println(hex.EncodeToHex(checkHash[:]))
+	println(k.Address().String())
+	println(invokerAddr.String())
 
 	invokeTx, err := invokerContract.Txn("invoke", is, it)
 	require.NoError(t, err)
@@ -166,153 +173,4 @@ func TestBasicInvoker(t *testing.T) {
 	require.NoError(t, err)
 	_ = rcpt
 
-}
-
-type InvokerSignature struct {
-	R *big.Int `abi:"r"`
-	S *big.Int `abi:"s"`
-	V bool     `abi:"v"`
-}
-
-func (is *InvokerSignature) SignCommit(key ethgo.Key, commit []byte, invokerAddr ethgo.Address) (err error) {
-	var msg [64]byte
-
-	// EIP-3074 messages are of the form
-	// keccak256(type ++ invoker ++ commit)
-	// TODO: this code will probably need to exist elsewhere - e.g. AUTH opcode
-	msg[0] = 0x03
-	copy(msg[13:33], invokerAddr.Bytes())
-	copy(msg[33:], commit)
-
-	var sig []byte
-	if sig, err = key.Sign(ethgo.Keccak256(msg[:])); err != nil {
-		return
-	}
-
-	is.R = new(big.Int).SetBytes(sig[0:32])
-	is.S = new(big.Int).SetBytes(sig[32:64])
-	if sig[64] == 0 {
-		is.V = false
-	} else {
-		is.V = true
-	}
-
-	return
-}
-
-type InvokerTransaction struct {
-	From     types.Address        `abi:"from"`
-	Nonce    *big.Int             `abi:"nonce"`
-	Payloads []TransactionPayload `abi:"payloads"`
-}
-
-// return keccak256(abi.encode(TRANSACTION_TYPE, transaction.from, transaction.nonce, hashPayloads(transaction.payloads)));
-
-type encTransaction struct {
-	TypeHash     [32]byte      `abi:"typeHash"`
-	From         types.Address `abi:"from"`
-	Nonce        *big.Int      `abi:"nonce"`
-	PayloadsHash [32]byte      `abi:"payloadsHash"`
-}
-
-var transactionType = ethgo.Keccak256([]byte("Transaction(address from,uint256 nonce,TransactionPayload[] payloads)TransactionPayload(address to,uint256 value,uint256 gasLimit,bytes data)"))
-
-func (it InvokerTransaction) InvokerCommit(domainSeparator []byte) (c []byte, err error) {
-	values := []byte{0x19, 0x01}
-	values = append(values, domainSeparator...)
-	if c, err = it.InvokerHash(); err != nil {
-		return
-	}
-	values = append(values, c...)
-	c = ethgo.Keccak256(values)
-	return
-}
-
-func (it InvokerTransaction) InvokerHash() (h []byte, err error) {
-
-	var t *abi.Type
-	if t, err = abi.NewType("tuple(bytes32 typeHash, address from, uint256 nonce, bytes32 payloadsHash)"); err != nil {
-		return
-	}
-
-	enc := encTransaction{
-		From:  it.From,
-		Nonce: it.Nonce,
-	}
-	copy(enc.TypeHash[:], transactionType)
-	tps := TransactionPayloads(it.Payloads)
-	if h, err = tps.InvokerHash(); err != nil {
-		return
-	}
-	copy(enc.PayloadsHash[:], h)
-
-	var encBytes []byte
-	if encBytes, err = abi.Encode(&enc, t); err != nil {
-		return
-	}
-
-	h = ethgo.Keccak256(encBytes)
-	return
-
-}
-
-type TransactionPayloads []TransactionPayload
-
-func (tps TransactionPayloads) InvokerHash() (h []byte, err error) {
-	var values []byte
-	for i := range tps {
-		if h, err = tps[i].InvokerHash(); err != nil {
-			return
-		}
-		values = append(values, h...)
-	}
-	// return keccak256(abi.encodePacked(values));
-	h = ethgo.Keccak256(values)
-	return
-}
-
-func (tps TransactionPayloads) Payloads() []TransactionPayload {
-	return tps
-}
-
-type TransactionPayload struct {
-	To       types.Address `abi:"to"`
-	Value    *big.Int      `abi:"value"`
-	GasLimit *big.Int      `abi:"gasLimit"`
-	Data     []byte        `abi:"data"`
-}
-
-type encPayload struct {
-	TypeHash [32]byte      `abi:"typeHash"`
-	To       types.Address `abi:"to"`
-	Value    *big.Int      `abi:"value"`
-	GasLimit *big.Int      `abi:"gasLimit"`
-	DataHash [32]byte      `abi:"dataHash"`
-}
-
-// from AccountAbstractionInvoker.sol: TRANSACTION_PAYLOAD_TYPE
-var transactionPayloadType = ethgo.Keccak256([]byte("TransactionPayload(address to,uint256 value,uint256 gasLimit,bytes data)"))
-
-func (tp TransactionPayload) InvokerHash() (h []byte, err error) {
-
-	var t *abi.Type
-	if t, err = abi.NewType("tuple(bytes32 typeHash, address to, uint256 value, uint256 gasLimit, bytes32 dataHash)"); err != nil {
-		return
-	}
-
-	enc := encPayload{
-		To:       tp.To,
-		Value:    tp.Value,
-		GasLimit: tp.GasLimit,
-	}
-	copy(enc.TypeHash[:], transactionPayloadType)
-	copy(enc.DataHash[:], ethgo.Keccak256(tp.Data))
-
-	var encBytes []byte
-	if encBytes, err = abi.Encode(&enc, t); err != nil {
-		return
-	}
-
-	h = ethgo.Keccak256(encBytes)
-	return
 }
