+++ state/runtime/evm/evm_test.go
@@ -1,6 +1,8 @@
 package evm
 
 import (
+	"bytes"
+	"github.com/0xPolygon/polygon-edge/helper/hex"
 	"math/big"
 	"testing"
 
@@ -11,12 +13,12 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func newMockContract(value *big.Int, gas uint64, code []byte) *runtime.Contract {
+func newMockContract(value *big.Int, gas uint64, code []byte, to types.Address) *runtime.Contract {
 	return runtime.NewContract(
 		1,
 		types.ZeroAddress,
 		types.ZeroAddress,
-		types.ZeroAddress,
+		to,
 		value,
 		gas,
 		code,
@@ -101,11 +103,31 @@ func (m *mockHost) GetRefund() uint64 {
 func TestRun(t *testing.T) {
 	t.Parallel()
 
+	// for AUTH test ...
+	// r: 0x8b1b4920a872ff8fdd66a35ef370a9d9113af4234b91ab0087c51e8362287073
+	// s: 0x68da649ea8d0444671dc2e3ed9b40ab9b0eee08d745a5527bdb92f0dff746d9c
+	// v: false
+	// commit: 0x6ca2b29e76ec69ab132c23acdafc5650de9f0ee2aa6ada70031962b37e24b026
+	// user: 0x3D09c91F44C87C30901dDB742D99f168F5AEEf01
+	// invoker: 0xC66298c7a6aDE36b928d6e9598Af7804611AbDC0
+
+	authCode := new(bytes.Buffer)
+	authCode.WriteByte(PUSH32)
+	authCode.Write(hex.MustDecodeHex("0x68da649ea8d0444671dc2e3ed9b40ab9b0eee08d745a5527bdb92f0dff746d9c"))
+	authCode.WriteByte(PUSH32)
+	authCode.Write(hex.MustDecodeHex("0x8b1b4920a872ff8fdd66a35ef370a9d9113af4234b91ab0087c51e8362287073"))
+	authCode.WriteByte(PUSH1)
+	authCode.WriteByte(0x00)
+	authCode.WriteByte(PUSH32)
+	authCode.Write(hex.MustDecodeHex("0x6ca2b29e76ec69ab132c23acdafc5650de9f0ee2aa6ada70031962b37e24b026"))
+	authCode.WriteByte(AUTH)
+
 	tests := []struct {
 		name     string
 		value    *big.Int
 		gas      uint64
 		code     []byte
+		to       types.Address
 		config   *chain.ForksInTime
 		expected *runtime.ExecutionResult
 	}{
@@ -164,6 +186,22 @@ func TestRun(t *testing.T) {
 				Err:     errRevert,
 			},
 		},
+		{
+			name:  "should succeed and return value of auth signer",
+			value: big.NewInt(0),
+			gas:   3100 + 4*3,
+			code:  authCode.Bytes(),
+			to:    types.StringToAddress("0xC66298c7a6aDE36b928d6e9598Af7804611AbDC0"),
+			config: &chain.ForksInTime{
+				Byzantium: true,
+			},
+			expected: &runtime.ExecutionResult{
+				ReturnValue: hex.MustDecodeHex("0x0000000000000000000000003D09c91F44C87C30901dDB742D99f168F5AEEf01"),
+				GasUsed:     3100 + 4*3,
+				GasLeft:     0,
+				Err:         nil,
+			},
+		},
 	}
 
 	for _, tt := range tests {
@@ -172,7 +210,7 @@ func TestRun(t *testing.T) {
 			t.Parallel()
 
 			evm := NewEVM()
-			contract := newMockContract(tt.value, tt.gas, tt.code)
+			contract := newMockContract(tt.value, tt.gas, tt.code, tt.to)
 			host := &mockHost{}
 			config := tt.config
 			if config == nil {
@@ -340,7 +378,7 @@ func TestRunWithTracer(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 
-			contract := newMockContract(tt.value, tt.gas, tt.code)
+			contract := newMockContract(tt.value, tt.gas, tt.code, types.ZeroAddress)
 			contract.Address = contractAddress
 			tracer := &mockTracer{}
 			host := &mockHost{
